Work around typing issues.
