+++ pyproject.toml
@@ -13,7 +13,7 @@ module = [
     "breezy.*",
     "buildlog_consultant.*",
     "lz4.*",
-    "ruamel.yaml.*",
+    "ruamel.*",
     "semver.*",
     "requirements.*",
     "setuptools.config.setupcfg",
